Use Camera instead of std::map to represent a set of Detector's.
In this module, a set of cameraGeom::Detector's in a camera is
represented by "CcdSet" class. CcdSet was previously just an alias
of std::map. Recently, they require cameraGeom::Camera to be present
if we want to change orientation of a cameraGeom::Detector. To cope
with this problem, we now directly use cameraGeom::Camera in place of
std::map. "CcdSet" class is now an alias of cameraGeom::Camera.
Because cameraGeom::Camera is immutable and is const-qualified, the
source code needs quite a little modification.